several fixes in the codebase, mostly in signed/unsigned checkings.

+++ lib/ext_max_record.h
@@ -3,5 +3,5 @@
  */
 int _gnutls_mre_num2record( int num);
 int _gnutls_mre_record2num( uint16 record_size);
-int _gnutls_max_record_recv_params( gnutls_session session, const opaque* data, int data_size);
-int _gnutls_max_record_send_params( gnutls_session session, opaque* data, int);
+int _gnutls_max_record_recv_params( gnutls_session session, const opaque* data, size_t data_size);
+int _gnutls_max_record_send_params( gnutls_session session, opaque* data, size_t);
