+++ lib/gnutls_buffers.h
@@ -18,9 +18,9 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA
  */
 
-int _gnutls_record_buffer_put(ContentType type, gnutls_session session, char *data, int length);
+int _gnutls_record_buffer_put(ContentType type, gnutls_session session, char *data, size_t length);
 int _gnutls_record_buffer_get_size(ContentType type, gnutls_session session);
-int _gnutls_record_buffer_get(ContentType type, gnutls_session session, char *data, int length);
+int _gnutls_record_buffer_get(ContentType type, gnutls_session session, char *data, size_t length);
 ssize_t _gnutls_io_read_buffered( gnutls_session, opaque **iptr, size_t n, ContentType);
 void _gnutls_io_clear_read_buffer( gnutls_session);
 int _gnutls_io_clear_peeked_data( gnutls_session session);
@@ -30,11 +30,11 @@ ssize_t _gnutls_io_write_buffered2( gnutls_session, const void *iptr, size_t n,
 	const void* iptr2, size_t n2);
 
 int _gnutls_handshake_buffer_get_size( gnutls_session session);
-int _gnutls_handshake_buffer_peek( gnutls_session session, char *data, int length);
-int _gnutls_handshake_buffer_put( gnutls_session session, char *data, int length);
+int _gnutls_handshake_buffer_peek( gnutls_session session, char *data, size_t length);
+int _gnutls_handshake_buffer_put( gnutls_session session, char *data, size_t length);
 int _gnutls_handshake_buffer_clear( gnutls_session session);
 int _gnutls_handshake_buffer_empty( gnutls_session session);
-int _gnutls_handshake_buffer_get_ptr( gnutls_session session, char **data_ptr, int *length);
+int _gnutls_handshake_buffer_get_ptr( gnutls_session session, char **data_ptr, size_t *length);
 
 #define _gnutls_handshake_io_buffer_clear( session) \
         gnutls_free( session->internals.handshake_send_buffer.data); \
